pyproject.toml
[project]
name = "saccade"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "ultralytics",
    "numpy",
    "opencv-python",
    "chromadb",
    "redis",
    "torch==2.11.0",
    "torchvision==0.26.0",
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "ruff",
    "pytest",
    "httpx",
    "nvidia-ml-py",
    "lap>=0.5.13",
    "mypy>=1.20.0",
    "pyyaml>=6.0.3",
    "anyio>=4.13.0",
    "pytest-asyncio>=1.3.0",
    "pytest-anyio>=0.0.0",
    "types-pyyaml>=6.0.12.20260408",
    "nvidia-dali-cuda120>=2.0.0",
    "pygobject==3.48.2",
    "timm>=1.0.26",
    "onnxscript>=0.6.2",
    "transformers>=4.48.0",
    "accelerate>=1.13.0",
    "optimum[onnxruntime-gpu]>=2.1.0",
    "einops>=0.8.2",
    "onnxruntime-gpu>=1.24.4",
    "tensorrt-cu12==10.16.1.11",
    "pybind11>=3.0.3",
    "pytest-cov>=6.0.0",
    "optuna>=4.8.0",
    "optuna-dashboard>=0.20.0",
    "motmetrics>=1.4.0",
    "llama-index>=0.14.21",
    "llama-index-vector-stores-chroma>=0.5.5",
]

[tool.uv]
override-dependencies = [
    "onnxruntime==1.24.4", # Force override to avoid multiple versions if needed, or use a dummy
]

[tool.mypy]
strict = true
ignore_missing_imports = true
exclude = ["tests/"]

[tool.ruff]
exclude = ["tests/benchmarks/archive"]

[tool.ruff.lint]
ignore = ["E701"] # Allow multiple statements on one line for now to avoid tedious formatting fixes in benchmarks

[tool.pytest.ini_options]
addopts = "--cov=perception --cov=cognition --cov=media --cov=storage --cov=pipeline --cov-report=term-missing"
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "strict"
